Change statements and bot functionality

+++ commands/Moderation/deletechannel.js
@@ -28,22 +28,26 @@ class DeleteChannelCommand extends Command {
 	
 	async run(bot, message, args, flags) {
 		let channel = args[0];
-		let code = Math.floor(Math.random() * 100000).toString();
-		if (code.length < 5) {
-			while (code.length < 5) code = "0" + code;
+
+		let cmdErr;
+		if (Number(new Date()) - channel.createdTimestamp > 1000 * 86400 * 180) {
+			let code = Math.floor(Math.random() * 100000).toString();
+			if (code.length < 5) {
+				while (code.length < 5) code = "0" + code;
+			}
+			message.channel.send(`You are about to delete the channel **${channel.name}**, which is more than 180 days old. Type \`${code}\` to proceed. This operation will time out in 30 seconds.`)
+			await message.channel.awaitMessages(msg => msg.author.id == message.author.id && msg.content == code, {
+				max: 1,
+				time: 30000,
+				errors: ["time"]
+			})
+			.catch(() => {cmdErr = true;})
 		}
-		message.channel.send(`You are about to delete the channel **${channel.name}**. Type \`${code}\` to proceed. This operation will time out in 30 seconds.`);
-		message.channel.awaitMessages(msg => msg.author.id == message.author.id && msg.content == code, {
-			max: 1,
-			time: 30000,
-			errors: ["time"]
-		})
-		.then(() => {
-			channel.delete()
-			.then(() => message.channel.send(`✅ The channel **${channel.name}** has been deleted.`))
-			.catch(() => message.channel.send("Could not delete the channel."))
-		})
-		.catch(() => message.channel.send("Operation canceled."))
+		if (cmdErr) return message.channel.send("Operation has timed out.");
+
+		channel.delete()
+		.then(() => message.channel.send(`✅ The channel **${channel.name}** has been deleted.`))
+		.catch(() => message.channel.send("Could not delete the channel."))
 	}
 }
